refactor(app): split history job drawer
